requirements.txt
# Core dependencies for Context Hive Hub tools
pyyaml==6.0.1

# Optional: for service implementation examples
fastapi==0.104.1
uvicorn==0.24.0

# Optional: for testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
